+++ awesome-react-app/public/common.css
@@ -42,9 +42,7 @@ main {
   padding: 2rem;
   display: flex;
   flex: 1;
-  justify-content: center;
   flex-direction: column;
-  align-items: center;
 }
 
 button {
