+++ awesome-react-app/src/components/App.css
@@ -21,6 +21,5 @@ input {
 }
 
 #root {
-  margin: 30px 0;
-  width: 1000px
+  width: 100%
 }
