+++ awesome-react-app/src/components/Modal/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const withModal = (komp: React.ReactNode) => {
+
+    
+    return null
+}
